+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Improvements:
     * runtime: ON WINDOW NEWSIZE CALL, ON WINDOW REFRESH CALL instructions added
     * runtime: GadToolsSupport module added
     * runtime: TEXTEXTEND statement added
+    * runtime: WINDOW() function: retrieve current screen's visible width and height (#49)
     * compiler: support for variadic function argument passing
     * compiler: NEW operator added
     * compiler: EXTENDS, BASE, IMPLEMENTS keywords added (OOP inheritance)